Add files via upload
The temperature converter app is a web-based project built using HTML, CSS, and JavaScript that allows users to convert temperatures between Celsius and Fahrenheit units effortlessly. Users can enter a value in one unit, and the app instantly displays the converted temperature in the other unit, providing a simple and user-friendly experience.